chore: moved singbox to services
