+++ docker-compose.yml
@@ -1,13 +1,12 @@
 services:
   junction:
-    build:
-      context: .
-      dockerfile: Dockerfile
-      target: standalone
+    image: ghcr.io/fmotalleb/junction:0.6.1-vless
     ports:
       - 7892:6980
       - 80:80
       - 443:443
+    environment:
+      - ZAPLOG_LEVEL=debug
     env_file:
       - .env
     networks:
